+++ src/main/java/winner/util/CheckUtil.java
@@ -1,4 +1,4 @@
-package winner;
+package winner.util;
 
 import java.util.regex.Pattern;
